requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
connexion==2.5.1
swagger-ui-bundle==0.0.6
dateutils==0.6.7
pykafka==2.8.0
python-dateutil==2.8.1
requests==2.22.0